# docs/api/openapi/storefront.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/storefront/cart": {
      "post": {
        "description": "Returns a one-time cart token. No shopper account is involved — whoever holds the token holds the cart.",
        "operationId": "CartController_create",
        "parameters": [
          {
            "name": "X-Merchant-Slug",
            "in": "header",
            "description": "Slug of the storefront being browsed",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-store"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCartDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedCartDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed X-Merchant-Slug header"
          },
          "401": {
            "description": "Missing or invalid cart token"
          },
          "404": {
            "description": "Storefront not found"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {}
        ],
        "summary": "Create a cart",
        "tags": [
          "Cart"
        ]
      }
    },
    "/storefront/cart/mine": {
      "get": {
        "description": "The one cart lookup that needs no X-Cart-Token, so a shopper who signs in on a new device can pick up where they left off. Requires a bearer token.",
        "operationId": "CartController_findMine",
        "parameters": [
          {
            "name": "X-Merchant-Slug",
            "in": "header",
            "description": "Slug of the storefront being browsed",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-store"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed X-Merchant-Slug header"
          },
          "401": {
            "description": "Missing or invalid cart token"
          },
          "404": {
            "description": "No active cart for this account"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {}
        ],
        "summary": "The signed-in shopper's current cart",
        "tags": [
          "Cart"
        ]
      }
    },
    "/storefront/cart/{cartId}": {
      "get": {
        "description": "Accepts either the X-Cart-Token or a bearer token for the shopper who owns the cart.",
        "operationId": "CartController_findOne",
        "parameters": [
          {
            "name": "X-Merchant-Slug",
            "in": "header",
            "description": "Slug of the storefront being browsed",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-store"
            }
          },
          {
            "name": "cartId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Cart-Token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed X-Merchant-Slug header"
          },
          "401": {
            "description": "Missing or invalid cart token"
          },
          "404": {
            "description": "Cart not found"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {}
        ],
        "summary": "Get a cart, re-priced at current catalog prices",
        "tags": [
          "Cart"
        ]
      }
    },
    "/storefront/cart/{cartId}/items": {
      "post": {
        "operationId": "CartController_addItem",
        "parameters": [
          {
            "name": "X-Merchant-Slug",
            "in": "header",
            "description": "Slug of the storefront being browsed",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-store"
            }
          },
          {
            "name": "cartId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Cart-Token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartItemInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed X-Merchant-Slug header"
          },
          "401": {
            "description": "Missing or invalid cart token"
          },
          "404": {
            "description": "Storefront not found"
          },
          "409": {
            "description": "Product is not purchasable here"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {}
        ],
        "summary": "Add an item, merging onto the matching line if present",
        "tags": [
          "Cart"
        ]
      },
      "delete": {
        "operationId": "CartController_clear",
        "parameters": [
          {
            "name": "X-Merchant-Slug",
            "in": "header",
            "description": "Slug of the storefront being browsed",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-store"
            }
          },
          {
            "name": "cartId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Cart-Token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed X-Merchant-Slug header"
          },
          "401": {
            "description": "Missing or invalid cart token"
          },
          "404": {
            "description": "Storefront not found"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {}
        ],
        "summary": "Empty the cart",
        "tags": [
          "Cart"
        ]
      }
    },
    "/storefront/cart/{cartId}/items/{itemId}": {
      "patch": {
        "operationId": "CartController_updateItem",
        "parameters": [
          {
            "name": "X-Merchant-Slug",
            "in": "header",
            "description": "Slug of the storefront being browsed",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-store"
            }
          },
          {
            "name": "cartId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Cart-Token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCartItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed X-Merchant-Slug header"
          },
          "401": {
            "description": "Missing or invalid cart token"
          },
          "404": {
            "description": "Cart item not found"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {}
        ],
        "summary": "Set a line quantity; zero removes the line",
        "tags": [
          "Cart"
        ]
      },
      "delete": {
        "operationId": "CartController_removeItem",
        "parameters": [
          {
            "name": "X-Merchant-Slug",
            "in": "header",
            "description": "Slug of the storefront being browsed",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-store"
            }
          },
          {
            "name": "cartId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Cart-Token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed X-Merchant-Slug header"
          },
          "401": {
            "description": "Missing or invalid cart token"
          },
          "404": {
            "description": "Cart item not found"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {}
        ],
        "summary": "Remove a line",
        "tags": [
          "Cart"
        ]
      }
    },
    "/storefront/cart/{cartId}/contact": {
      "patch": {
        "operationId": "CartController_updateContact",
        "parameters": [
          {
            "name": "X-Merchant-Slug",
            "in": "header",
            "description": "Slug of the storefront being browsed",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-store"
            }
          },
          {
            "name": "cartId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Cart-Token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCartContactDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed X-Merchant-Slug header"
          },
          "401": {
            "description": "Missing or invalid cart token"
          },
          "404": {
            "description": "Storefront not found"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {}
        ],
        "summary": "Set the shopper name, email, phone or order note",
        "tags": [
          "Cart"
        ]
      }
    },
...
    "/storefront/cart/{cartId}/checkout": {
      "post": {
        "description": "Returns a checkout session plus its X-Checkout-Token; the cart is then CONVERTED.",
        "operationId": "CartController_checkout",
        "parameters": [
          {
            "name": "X-Merchant-Slug",
            "in": "header",
            "description": "Slug of the storefront being browsed",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-store"
            }
          },
          {
            "name": "cartId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Cart-Token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutCartDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutSessionDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed X-Merchant-Slug header"
          },
          "401": {
            "description": "Missing or invalid cart token"
          },
          "404": {
            "description": "Storefront not found"
          },
          "409": {
            "description": "Cart is empty, has no contact details, or its delivery selection is no longer valid"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {}
        ],
        "summary": "Convert the cart into a checkout session and reserve stock",
        "tags": [
          "Cart"
        ]
      }
    },
...
  },
  "info": {
    "title": "Storefront API",
    "description": "Storefront browsing, carts, shopper addresses (guest via cart token, or signed-in via bearer token), delivery options and tracking, checkout sessions, payment initiation, and payment status polling",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "checkout-token": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Checkout-Token"
      },
      "cart-token": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Cart-Token"
      }
    },
    "schemas": {
...
      "CartItemInputDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "variantId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "number",
            "minimum": 1,
            "maximum": 100
          },
          "note": {
            "type": "string",
            "maxLength": 200
          }
        },
        "required": [
          "productId",
          "quantity"
        ]
      },
      "CreateCartDto": {
        "type": "object",
        "properties": {
          "sourceChannel": {
            "type": "string",
            "enum": [
              "POS",
              "WEBSITE",
              "FACEBOOK",
              "INSTAGRAM",
              "TIKTOK"
            ],
            "default": "WEBSITE",
            "description": "POS is rejected; POS sells through its own flow"
          },
          "items": {
            "description": "Optional seed lines, so a cart can be created pre-filled",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItemInputDto"
            }
          }
        }
      },
      "CartLineDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "object",
            "format": "uuid",
            "nullable": true,
            "description": "The cart item id, used to update or remove the line"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "variantId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "sku": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "unitPrice": {
            "type": "string",
            "example": "19.99"
          },
          "totalPrice": {
            "type": "string",
            "example": "39.98"
          },
          "note": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "productId",
          "sku",
          "name",
          "quantity",
          "unitPrice",
          "totalPrice"
        ]
      },
...
      "CreatedCartDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "CONVERTED",
              "MERGED",
              "ABANDONED",
              "EXPIRED"
            ]
          },
          "sourceChannel": {
            "type": "string",
            "enum": [
              "POS",
              "WEBSITE",
              "FACEBOOK",
              "INSTAGRAM",
              "TIKTOK"
            ]
          },
          "customerId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "ownerId": {
            "type": "object",
            "format": "uuid",
            "nullable": true,
            "description": "The signed-in shopper this cart belongs to, once claimed. Null for a guest cart."
          },
          "customerName": {
            "type": "object",
            "nullable": true
          },
          "customerEmail": {
            "type": "object",
            "nullable": true
          },
          "customerPhone": {
            "type": "object",
            "nullable": true
          },
          "note": {
            "type": "object",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "example": "USD"
          },
          "subtotalAmount": {
            "type": "string",
            "example": "39.98"
          },
          "shippingAmount": {
            "type": "string",
            "example": "2.50"
          },
          "totalAmount": {
            "type": "string",
            "example": "42.48"
          },
          "itemCount": {
            "type": "number",
            "example": 2
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartLineDto"
            }
          },
          "shippingAddress": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/AddressDto"
              }
            ]
          },
          "billingAddress": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/AddressDto"
              }
            ]
          },
          "delivery": {
            "nullable": true,
            "description": "The chosen delivery method, re-quoted at read time",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/DeliveryOptionDto"
              }
            ]
          },
          "checkoutSessionId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "cartToken": {
            "type": "string",
            "description": "Send as the X-Cart-Token header on every later request for this cart. Shown once."
          }
        },
        "required": [
          "id",
          "status",
          "sourceChannel",
          "currency",
          "subtotalAmount",
          "shippingAmount",
          "totalAmount",
          "itemCount",
          "items",
          "expiresAt",
          "cartToken"
        ]
      },
      "CartDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "CONVERTED",
              "MERGED",
              "ABANDONED",
              "EXPIRED"
            ]
          },
          "sourceChannel": {
            "type": "string",
            "enum": [
              "POS",
              "WEBSITE",
              "FACEBOOK",
              "INSTAGRAM",
              "TIKTOK"
            ]
          },
          "customerId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "ownerId": {
            "type": "object",
            "format": "uuid",
            "nullable": true,
            "description": "The signed-in shopper this cart belongs to, once claimed. Null for a guest cart."
          },
          "customerName": {
            "type": "object",
            "nullable": true
          },
          "customerEmail": {
            "type": "object",
            "nullable": true
          },
          "customerPhone": {
            "type": "object",
            "nullable": true
          },
          "note": {
            "type": "object",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "example": "USD"
          },
          "subtotalAmount": {
            "type": "string",
            "example": "39.98"
          },
          "shippingAmount": {
            "type": "string",
            "example": "2.50"
          },
          "totalAmount": {
            "type": "string",
            "example": "42.48"
          },
          "itemCount": {
            "type": "number",
            "example": 2
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartLineDto"
            }
          },
          "shippingAddress": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/AddressDto"
              }
            ]
          },
          "billingAddress": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/AddressDto"
              }
            ]
          },
          "delivery": {
            "nullable": true,
            "description": "The chosen delivery method, re-quoted at read time",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/DeliveryOptionDto"
              }
            ]
          },
          "checkoutSessionId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "sourceChannel",
          "currency",
          "subtotalAmount",
          "shippingAmount",
          "totalAmount",
          "itemCount",
          "items",
          "expiresAt"
        ]
      },
      "UpdateCartItemDto": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "number",
            "minimum": 0,
            "maximum": 100,
            "description": "Zero removes the line"
          },
          "note": {
            "type": "string",
            "maxLength": 200
          }
        },
        "required": [
          "quantity"
        ]
      },
      "UpdateCartContactDto": {
        "type": "object",
        "properties": {
          "customerName": {
            "type": "string",
            "maxLength": 120
          },
          "customerEmail": {
            "type": "string",
            "format": "email"
          },
          "customerPhone": {
            "type": "string",
            "maxLength": 30
          },
          "note": {
            "type": "string",
            "maxLength": 500
          }
        }
      },
...
      "CheckoutCartDto": {
        "type": "object",
        "properties": {
          "expiresInMinutes": {
            "type": "number",
            "default": 15,
            "minimum": 1,
            "maximum": 120
          }
        }
      },
...
    }
  }
}